examples/e2b-dev-sidecar/requirements.txt
aiohttp>=3.9
e2b-code-interpreter
python-dotenv
